data-platform/lambda/ingest/requirements.txt
jquants-api-client>=2.0.0
tenacity>=8.0.0
boto3>=1.34.0
pandas>=2.2.0
